Fin sprint 5 correction Crud Docente y Campus

+++ app/Http/Controllers/CampusController.php
@@ -1,11 +1,18 @@
 <?php namespace App\Http\Controllers;
-
+/* 
+ * Sistema de Gestion de Laboratorios - ESPE
+ *
+ * Author: Barrera Erick - LLamuca Andrea
+ * Revisado por: 
+ *
+*/
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Campus;
+use App\Laboratorio;
 
 use Illuminate\Http\Request;
-use Illuminate\Http\RedirectResponse;
+use DB;
 
 class CampusController extends Controller {
 
@@ -33,18 +40,24 @@ class CampusController extends Controller {
 	/**
 	 * Store a newly created resource in storage.
 	 *
-	 * @param  Request  $request
 	 * @return Response
 	 */
 	public function store(Request $request)
 	{
-		Campus::create([
-			'CAM_CODIGO' => $request['CAM_CODIGO'],
-			'CAM_NOMBRE' => $request['CAM_NOMBRE']
-		]);
-		return redirect('campus')
-			->with('title','Campus creada!')
-			->with('subtitle','Se ha creado correctamente el campus.');
+		$valida=Campus::where('CAM_NOMBRE',$request['CAM_NOMBRE'])->first();
+		if(empty($valida)){
+			Campus::create([
+				'CAM_CODIGO' => $request['CAM_CODIGO'],
+				'CAM_NOMBRE' => $request['CAM_NOMBRE']
+			]);
+			return redirect('campus')
+				->with('title','Campus creado!')
+				->with('subtitle','El registro de campus se ha realizado con éxito');
+		}else{
+            return view('campus.create')            	
+				->with('CAM_NOMBRE',$request->CAM_NOMBRE)
+            	->with('mensajes','El nombre del campus ya existe!, NO SE DEBE REPETIR');
+		}	
 	}
 
 	/**
@@ -67,23 +80,33 @@ class CampusController extends Controller {
 	public function edit($id)
 	{
 		$campus = Campus::find($id);
-		return view('campus.update', ['campus' => $campus]);
+		return view('campus.update', 
+			['campus' => $campus]);
 	}
 
 	/**
 	 * Update the specified resource in storage.
 	 *
-	 * @param  Request  $request
+	 * @param  int  $id
 	 * @return Response
 	 */
 	public function update(Request $request)
 	{
-		$campus = Campus::find($request['CAM_CODIGO']);
+		$campusValida=Campus::where('CAM_NOMBRE',$request['CAM_NOMBRE'])->first();
+		$campus = Campus::find($request['CAM_CODIGO']); 
+		if (count($campusValida) === 1){
+			if ($campusvalida->CAM_CODIGO === $campus->CAM_CODIGO){
+			}else{
+				return view('campus.update', ['campus' => $campus])            	
+					->with('CAM_NOMBRE',$request->CAM_NOMBRE)
+	            	->with('mensajes','El nombre del campus ya existe!, NO SE DEBE REPETIR');
+	        }
+		}  
 		$campus->fill($request->all());
 		$campus->save();
 		return redirect('campus')
 			->with('title','Campus actualizado!')
-			->with('subtitle','Se han actualizado correctamente los datos del campus.');
+			->with('subtitle','Se han actualizado correctamente los datos del campus.');   
 	}
 
 	/**
@@ -94,10 +117,18 @@ class CampusController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		Campus::destroy($id);
-		return redirect('campus')
-			->with('title','Campus eliminado!')
-			->with('subtitle','Se ha eliminado correctamente el campus.');
+		$validaLaboratorio = Laboratorio::where('CAM_CODIGO',$id)->first();
+		
+		if (count($validaLaboratorio) === 1) {
+			return redirect('campus')
+				->with('title', 'Campus NO Eliminado!')
+				->with('subtitle', 'El registro del campus no se a eliminado correctamente, el campus tiene registros relacionados.');
+		}else{
+			Campus::destroy($id);
+			return redirect('campus')
+				->with('title', 'Campus eliminado!')
+				->with('subtitle', 'La eliminación del campus se ha realizado con éxito.');
+		}
 	}
 
 }
